fix for GP-6584, set 'log.dir' location
* set fileName in log4j2.xml
* refactored classes as needed
-- created VisualizerLauncherGui class, from VisualizerLauncher
-- created AppDirUtil class, from FileUtil
