+++ src/main/java/org/genepattern/desktop/TaskInfo.java
@@ -61,7 +61,7 @@ public class TaskInfo {
         }
         
         // <app.dir>/taskLib/<task.name>_v<task.version>/
-        final File appDir=FileUtil.getAppDir();
+        final File appDir=AppDirUtil.getAppDir();
         taskInfo.libdir=new File(appDir, "taskLib/"+taskInfo.name+"_v"+taskInfo.version);
         return taskInfo;
     }
