+++ src/main/resources/log4j2.xml
@@ -4,7 +4,7 @@
         <Console name="Console" target="SYSTEM_OUT">
             <PatternLayout pattern="[%level] %msg%n" />
         </Console>
-        <File name="LogFile" fileName="logs/all.log" immediateFlush="true" append="true">
+        <File name="LogFile" fileName="${sys:log.dir:-logs}/all.log" immediateFlush="true" append="true">
             <PatternLayout pattern="%d{yyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
         </File>
     </Appenders>
